+++ CHANGELOG.md
@@ -29,15 +29,16 @@ This release contains upgrade of `sentry-android` dependency to major version 7.
 ### Fixes
 
 - Upload Debug Symbols Build Phase continues when `node` not found in `WITH_ENVIRONMENT` ([#3573](https://github.com/getsentry/sentry-react-native/pull/3573))
+- Fix `proguardUuid` loading on Android ([#3591](https://github.com/getsentry/sentry-react-native/pull/3591))
 
 ### Dependencies
 
 - Bump Android SDK from v6.34.0 to v7.3.0 ([#3434](https://github.com/getsentry/sentry-react-native/pull/3434))
   - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#730)
   - [diff](https://github.com/getsentry/sentry-java/compare/6.34.0...7.3.0)
-- Bump JavaScript SDK from v7.81.1 to v7.99.0 ([#3426](https://github.com/getsentry/sentry-react-native/pull/3426))
+- Bump JavaScript SDK from v7.81.1 to v7.100.1 ([#3426](https://github.com/getsentry/sentry-react-native/pull/3426), [#3589](https://github.com/getsentry/sentry-react-native/pull/3589))
   - [changelog](https://github.com/getsentry/sentry-javascript/blob/develop/CHANGELOG.md#7990)
-  - [diff](https://github.com/getsentry/sentry-javascript/compare/7.81.1...7.99.0)
+  - [diff](https://github.com/getsentry/sentry-javascript/compare/7.81.1...7.100.1)
 
 ## 5.18.0